[GREEN] Finish implementing Login Page component by refactoring the old ones.
